+++ packages/react/src/hooks/useResolvedTemplate.ts
@@ -1,7 +1,15 @@
 import { useEffect, useState } from "react";
-import { Template, Fetcher, ResolvedTemplate } from "@carrot-kpi/sdk";
+import {
+    Template,
+    Fetcher,
+    ResolvedTemplate,
+    getServiceURL,
+    Service,
+} from "@carrot-kpi/sdk";
 import { useNetwork } from "wagmi";
 import { useIPFSGatewayURL } from "./useIPFSGatewayURL";
+import { useProdMode } from "./useProdMode";
+import { usePreferDecentralization } from "./usePreferDecentralization";
 
 interface ResolvedTemplateParams {
     template?: Template;
@@ -13,6 +21,8 @@ export function useResolvedTemplate(params?: ResolvedTemplateParams): {
 } {
     const { chain } = useNetwork();
     const ipfsGatewayURL = useIPFSGatewayURL();
+    const preferDecentralization = usePreferDecentralization();
+    const prodMode = useProdMode();
 
     const [resolvedTemplate, setResolvedTemplate] =
         useState<ResolvedTemplate | null>(null);
@@ -26,6 +36,8 @@ export function useResolvedTemplate(params?: ResolvedTemplateParams): {
             try {
                 const resolved = await Fetcher.resolveTemplates({
                     ipfsGatewayURL,
+                    dataCDNURL: getServiceURL(Service.DATA_CDN, prodMode),
+                    preferDecentralization,
                     templates: [params.template],
                 });
                 if (resolved.length !== 1) return;
@@ -43,7 +55,13 @@ export function useResolvedTemplate(params?: ResolvedTemplateParams): {
         return () => {
             cancelled = true;
         };
-    }, [chain, ipfsGatewayURL, params?.template]);
+    }, [
+        chain,
+        ipfsGatewayURL,
+        prodMode,
+        preferDecentralization,
+        params?.template,
+    ]);
 
     return { loading, resolvedTemplate };
 }