+++ packages/react/src/hooks/useResolvedTemplates.ts
@@ -1,7 +1,15 @@
 import { useEffect, useState } from "react";
-import { Template, Fetcher, ResolvedTemplate } from "@carrot-kpi/sdk";
+import {
+    Template,
+    Fetcher,
+    ResolvedTemplate,
+    getServiceURL,
+    Service,
+} from "@carrot-kpi/sdk";
 import { useNetwork } from "wagmi";
 import { useIPFSGatewayURL } from "./useIPFSGatewayURL";
+import { usePreferDecentralization } from "./usePreferDecentralization";
+import { useProdMode } from "./useProdMode";
 
 interface ResolvedTemplatesParams {
     templates?: Template[];
@@ -13,6 +21,8 @@ export function useResolvedTemplates(params?: ResolvedTemplatesParams): {
 } {
     const { chain } = useNetwork();
     const ipfsGatewayURL = useIPFSGatewayURL();
+    const preferDecentralization = usePreferDecentralization();
+    const prodMode = useProdMode();
 
     const [resolvedTemplates, setResolvedTemplates] = useState<
         ResolvedTemplate[]
@@ -28,6 +38,8 @@ export function useResolvedTemplates(params?: ResolvedTemplatesParams): {
             try {
                 const resolved = await Fetcher.resolveTemplates({
                     ipfsGatewayURL,
+                    dataCDNURL: getServiceURL(Service.DATA_CDN, prodMode),
+                    preferDecentralization,
                     templates: params.templates,
                 });
                 if (resolved.length !== 1) return;
@@ -42,7 +54,13 @@ export function useResolvedTemplates(params?: ResolvedTemplatesParams): {
         return () => {
             cancelled = true;
         };
-    }, [chain, ipfsGatewayURL, params?.templates]);
+    }, [
+        chain,
+        ipfsGatewayURL,
+        preferDecentralization,
+        prodMode,
+        params?.templates,
+    ]);
 
     return { loading, resolvedTemplates };
 }