+++ packages/sdk/src/utils/services.ts
@@ -1,7 +1,7 @@
 export enum Service {
     STATIC_CDN,
-    IPFS_GATEWAY,
-    PINNING_PROXY,
+    DATA_CDN,
+    DATA_UPLOADER,
 }
 
 export interface ServiceURLs {
@@ -14,13 +14,13 @@ const SERVICE_URLS: Record<Service, ServiceURLs> = {
         staging: "https://static.staging.carrot.community",
         prod: "https://static.carrot.community",
     },
-    [Service.IPFS_GATEWAY]: {
-        staging: "https://gateway.api.staging.carrot.community",
-        prod: "https://gateway.api.carrot.community",
+    [Service.DATA_CDN]: {
+        staging: "https://data.staging.carrot.community",
+        prod: "https://data.carrot.community",
     },
-    [Service.PINNING_PROXY]: {
-        staging: "https://pinning-proxy.api.staging.carrot.community",
-        prod: "https://pinning-proxy.api.carrot.community",
+    [Service.DATA_UPLOADER]: {
+        staging: "https://data-uploader.api.staging.carrot.community",
+        prod: "https://data-uploader.api.carrot.community",
     },
 };